Problème avec FragmentController : gérer soucis compteur Gestion score ok, gestion affichage ok
